It looks like leaving out the `typeRoots` entirely fixes #196 - pioneer still starts and seems possible to navigate - a linked `joy-types` still builds - building `joy-types` in its own directory still works
The rationale for `typeRoots` was that prior to yarn workspaces in the
storage node, resolving types was a bit dicey. With workspaces and
typescript's documented behaviour, it seems better for packages to
try to resolve types in the typescript way.

See [@types, typeRoots and
types in the tsconfig.json section](https://www.typescriptlang.org/docs/handbook/tsconfig-json.html)