Add test, and fixtures, for reproducing update_category bug
I was basically using the same check I always do to see if the category path to the root is not already archived or deleted, the problem is this will always say YES when you are trying to unarchive/undelete a category. What I should be doing is skipping the category itself, and only check from its parent to the root.